refactor(Turn): 내부 static 클래스로 전환
