--- src/main/java/Turn.java
@@ -1,17 +1,0 @@
-import domain.piece.Team;
-
-public class Turn {
-    private Team team;
-
-    public Turn(Team team) {
-        this.team = team;
-    }
-
-    public void changeTurn() {
-        this.team = team.inverse();
-    }
-
-    public Team team() {
-        return team;
-    }
-}
